+++ pyproject.toml
@@ -35,6 +35,7 @@ markers =[
 python_files = "test_*.py *utils.py"
 
 [tool.mypy]
+python_version = "3.12"
 files = [
     "utils/",
     "tests/",
@@ -99,7 +100,6 @@ ignore = [
     "PLR0912",  # too-many-branches: TODO (17 occurences)
     "PLR0913",  # too-many-arguments: TODO (66 occurences)
     "PLR0915",  # too-many-statements: TODO (7 occurences)
-    "PLR2004",  # magic-value-comparison: TODO (674 occurences)
     "PTH118",   # os-path-join: TODO (59 occurences)
     "PTH123",   # builtin-open: TODO (117 occurences)
     "RUF012",   # mutable-class-default: TODO (111 occurences)
@@ -177,6 +177,7 @@ ignore = [
     "FIX003",    # line-contains-xxx: freedom of speech!
     "FIX004",    # line-contains-xxx: freedom of speech!
     "PLR1730",   # if-stmt-min-max: not clear that it makes the code easier to read
+    "PLR2004",   # magic-value-comparison: test code has lot of hardocded values
     "PT011",     # pytest-raises-too-broad: kind of painful to fix
     "RET506",    # superfluous-else-raise: requires a slightly higher cognitive effort to understand the code
     "RET507",    # superfluous-else-continue : requires a slightly higher cognitive effort to understand the code
@@ -226,6 +227,7 @@ ignore = [
     "RET505",   # superfluous-else-return: TODO
     "SIM108",   # if-else-block-instead-of-if-exp: TODO
     "SIM210",   # if-expr-with-true-false: TODO
+    "PLR2004",  # magic-value-comparison: TODO
 ]
 "utils/_features.py" = [
     "ANN001", # missing-type-function-argument: obvious decorators
@@ -276,6 +278,7 @@ ignore = [
     "FBT002",  # boolean-default-value-positional-argument: TODO
     "N803",    # invalid-argument-name: TODO
     "N806",    # non-lowercase-variable-in-function: TODO
+    "PLR2004", # magic-value-comparison: TODO
     "PLR2044", # empty-comment: TODO
     "PTH102",  # os-mkdir: TODO
     "RET505",  # superfluous-else-return: TODO
@@ -285,10 +288,15 @@ ignore = [
 "utils/parametric/*" = [
     "ANN001",   # missing-type-function-argument: TODO
     "ANN003",   # missing-type-kwargs: TODO
+    "PLR2004",  # magic-value-comparison: TODO
 ]
 "utils/scripts/ci_orchestrator/gitlab_exporter.py" = [ "ANN001" ]  # missing-type-function-argument: TODO
 "utils/scripts/junit_report.py" = ["ANN001"] # missing-type-function-argument: TODO
-"utils/scripts/ssi_wizards/aws_onboarding_wizard_utils.py" = ["ANN001"] # missing-type-function-argument: TODO
+"utils/scripts/ssi_wizards/aws_onboarding_wizard_utils.py" = [
+    "ANN001",   # missing-type-function-argument: TODO
+    "PLR2004",  # magic-value-comparison: TODO
+]
+     
 "utils/scripts/ci_orchestrators/gitlab_exporter.py" = ["ANN001"] # missing-type-function-argument: TODO
 "utils/scripts/ci_orchestrators/external_gitlab_pipeline.py" = ["ANN001"] # missing-type-function-argument: TODO
 "utils/scripts/*" = [
@@ -301,6 +309,7 @@ ignore = [
     "ANN201",   # missing-return-type-undocumented-public-function: TODO
     "ARG002",   # unused-method-argument: TODO
     "FBT002",   # boolean-default-value-positional-argument: TODO
+    "PLR2004",  # magic-value-comparison: TODO
     "TRY002",   # raise-vanilla-class: TODO
 ]
 "utils/virtual_machine/aws_provider.py" = [
@@ -321,6 +330,7 @@ ignore = [
     "F401",    # unused-import: TODO
     "F841",    # unused-variable: TODO
     "TRY002",  # raise-vanilla-class: TODO
+    "PLR2004", # magic-value-comparison: TODO
 ]
 "utils/virtual_machine/krunvm_provider.py" = [
     "ANN001",  # missing-type-function-argument: TODO
