+++ utils/_context/_scenarios/external_processing.py
@@ -52,7 +52,7 @@ class ExternalProcessingScenario(DockerScenario):
         interfaces.agent.configure(self.host_log_folder, replay=self.replay)
 
     def _start_interfaces_watchdog(self):
-        super()._start_interfaces_watchdog([interfaces.library, interfaces.agent])
+        super().start_interfaces_watchdog([interfaces.library, interfaces.agent])
 
     def _wait_for_app_readiness(self):
         logger.debug("Wait for app readiness")
